backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
google-generativeai==0.3.1
httpx==0.25.1
aiohttp==3.9.1
requests==2.31.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
python-nmap==0.7.1
dnspython==2.4.2
markdown==3.5.1
jinja2==3.1.2
colorama==0.4.6
rich==13.7.0
